Fix drag-drop for buttonGroups
Previously a boolean model would be dragged when Ctrl-Dragging on a
button belonging to a button group.  This fix causes the integer-model
of the group to be dragged instead.

The original behavior would allow the user to toggle multiple values
simultaneously by using several automation tracks.  Plus, I would think
this new behavior is more convenient.

git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1968 0778d3d1-df1d-0410-868b-ea421aaaa00d

Conflicts:
	ChangeLog.old
	src/gui/widgets/automatable_button.cpp