ResourceTreeModel: dynamic icons for plugin specific resources + caching
Added dynamic generation of icons for plugin specific resources by
drawing the logo of the according plugin onto an empty-sheet-icon.
After generation, cache the result so this expensive operation only
has to be done once.

Furthermore also cache pixmaps for individual ResourceItem types for
not having to load and scale pixmaps for each item.

Signed-off-by: Tobias Doerffel <tobias.doerffel@gmail.com>