ZynAddSubFX: new QtXmlWrapper, a QtXml-based XML backend for ZASF
ZynAddSubFX comes with a default XMLwrapper class which uses mxml
library as backend. It is however does not make much sense to ship mxml
as an additional XML framework. With the new QtXmlWrapper class we can
drop all mxml source code.

Saving and loading settings/presets works well so far, but there still
might be bugs in this implementation.