ZynAddSubFX/FLTK: The last freetype fix, I promise.
Using FREETYPE_INCLUDE_DIRS turned out to break the build as it is a list consisting of two paths.
During configuring, the list is then concatenated to "path1 path2". This makes cmake to think of the second path as the directory to configure.
Fixed by manually concatenating the strings with a ";" character.