InstrumentTrack: set range of FX channel model to infinite when loading settings
When loading a project, instrument tracks usually are instantiated/loaded
before the FX mixer settings are loaded which results in an empty FX mixer
and thus 0 FX channels. The actual FX channel value for the instrument track
would be lost that way. Therefore set a big value which is being corrected
later by the FxMixerView for all instruments.