Remove inline function recursion for Apple
q_sort is an inline function which recurses upon itself.  Some compilers such as MSVC limit this which can be overridden with a parameter.  Since I could find no such parameter for clang, this #IFDEF allows the sifter plugin to compile successfully.

If this approach has negative side effects, it should only affect apple.  If there's a better approach to this, please reject the pull request and submit your own.