Make time widget respect tempo automation Also I changed the way we keep track of time in song.cpp: instead of counting milliseconds there, we count frames, and then convert to milliseconds when needed. This is much more accurate and it makes it simple to take tempo in account with the new globalAutomationValueAt() method.
From this, it should be a simple step to make sample tracks playable from any point in the song: send a signal whenever play position is changed, have sampletracks receive the signal and recalculate their own playback position...

A caveat: tempo automations in bb-tracks do not work yet, but that's a pretty rare corner case so it should be fine.
Anyway they can be made to work with some additional code in AutomatableModel.