Automation pattern flip
Ok. Retrying with a couple of fixes:
*All lowercase art names
*No thumbs.db
*"Whats This" text added in Automation Editor
*update() after flip from context menu
*emit dataChanged() after flip
I can't compile to test that it's working as I run out of space whenever
I try, but I guess we'll see if the Travis build passes. (If it does,
all the features should still work the way they did in th last pull
request)