Two new presets
The clap is nothing more than a dry clap. As of now there is a clap with
decay or something. I used two evenings to make the clap I wanted so I
figured other people maybe wanted a dry clap too. Of course it is
probably I who suck at adjusting knobs correctly, but I reckon there are
more of us. :worried:

I probably didn't spend too much time on the kick, but as of now there
is only a trapkick in LMMS (kick with long delay), and though we agree
on that it is easy to make kicks in Kicker, it is even easier to use
existing presets and adjust a preexisting preset. So I missed a short
kick in LMMS, and so I made one.