fix for some minor issues
    const NumKeys (note.h) was used differently, sometimes including c9,
sometimes not, resulting in some bugs / strange behaviour around the
upper border of the piano roll editor.
    in the piano roll editor it was possible to move notes out of the
window, so I went through all the places, where NumKeys was used and
replaced it.
    Now there is: NumKeys (actual number of keys, 109 = 9 octaves * 12
tones + c9), MaxKey (108 = c9) and MinKey (0, replaced where suited)

    the piano roll editor didn't react properly on resize of the note
edit area. you could make it smaller when scrolled all up and new notes
would appear above c9.
    Now: void recalculateSize() (PianoRoll.h) is called from
PianoRoll::mouseMoveEvent() and PianoRoll::paintEvent()

    there was no possibility to change the volume of an multiple notes.
    Now: void addVolume(int amount) (PianoRoll.h) works like
shiftSemiTone().
    Shift + PageUp/Down changes the volume of a selection by 1,
    Control + PageUp/Down by 10.

    shiftSemiTone (PianoRoll.h) moved all notes as far as possible,
resulting in a "crunched" melody, when hitting the border.
    Now: when moving up/down a selection with shift+up/down or control
piano
    I changed PianoRoll::keyPressEvent() accordingly

    I was missing keyboard shortcuts, when clicking on the cogwheel of a
track in the song editor, like there are when right-clicking on an
effect in the effects chain. This is handy for removing multiple tracks,
for example, or especially when you have to change the midi input device
often between tracks.
    Now: added &, where useful, so now menu shortcuts work